LayoutTests/html5lib/runner-expected-html5.txt

CONSOLE MESSAGE: line 1: SyntaxError: Parse error
CONSOLE MESSAGE: line 1: SyntaxError: Parse error
CONSOLE MESSAGE: line 2: PASS
CONSOLE MESSAGE: line 2: FOO<span>BAR</span>BAZ
Content-Type: text/plain
resources/tests1.dat:
30
32
52
53
78
79
80
103

Test 30 of 113 in resources/tests1.dat failed. Input:
<a><table><td><a><table></table><a></tr><a></table><b>X</b>C<a>Y
Got:
| <html>
|   <head>
|   <body>
|     <a>
|       <a>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <a>
|                 <table>
|               <a>
|       <a>
|         <b>
|           "X"
|         "C"
|       <a>
|         "Y"
Expected:
| <html>
|   <head>
|   <body>
|     <a>
|       <a>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <a>
|                 <table>
|               <a>
|     <a>
|       <b>
|         "X"
|       "C"
|     <a>
|       "Y"

Test 32 of 113 in resources/tests1.dat failed. Input:
<!-----><font><div>hello<table>excite!<b>me!<th><i>please!</tr><!--X-->
Got:
| <!-- - -->
| <html>
|   <head>
|   <body>
|     <font>
|       <div>
|         "hello"
|         <b>
|           "me!"
|         <table>
|           "excite!"
|           <tbody>
|             <tr>
|               <th>
|                 <i>
|                   "please!"
|             <!-- X -->
Expected:
| <!-- - -->
| <html>
|   <head>
|   <body>
|     <font>
|       <div>
|         "helloexcite!"
|         <b>
|           "me!"
|         <table>
|           <tbody>
|             <tr>
|               <th>
|                 <i>
|                   "please!"
|             <!-- X -->

Test 52 of 113 in resources/tests1.dat failed. Input:
<p id=a><b><p id=b></b>TEST
Got:
| <html>
|   <head>
|   <body>
|     <p>
|       id="a"
|       <b>
|     <p>
|       id="b"
|       <b>
|         "TEST"
Expected:
| <html>
|   <head>
|   <body>
|     <p>
|       id="a"
|       <b>
|     <p>
|       id="b"
|       "TEST"

Test 53 of 113 in resources/tests1.dat failed. Input:
<b id=a><p><b id=b></p></b>TEST
Got:
| <html>
|   <head>
|   <body>
|     <b>
|       id="a"
|       <p>
|         <b>
|           id="b"
|       <b>
|         "TEST"
Expected:
| <html>
|   <head>
|   <body>
|     <b>
|       id="a"
|       <p>
|         <b>
|           id="b"
|       "TEST"

Test 78 of 113 in resources/tests1.dat failed. Input:
<a href="blah">aba<table><a href="foo">br<tr><td></td></tr>x</table>aoe
Got:
| <html>
|   <head>
|   <body>
|     <a>
|       href="blah"
|       "aba"
|       <a>
|         href="foo"
|         "br"
|       <table>
|         <tbody>
|           <tr>
|             <td>
|           "x"
|     "aoe"
Expected:
| <html>
|   <head>
|   <body>
|     <a>
|       href="blah"
|       "aba"
|       <a>
|         href="foo"
|         "br"
|       <a>
|         href="foo"
|         "x"
|       <table>
|         <tbody>
|           <tr>
|             <td>
|     <a>
|       href="foo"
|       "aoe"

Test 79 of 113 in resources/tests1.dat failed. Input:
<a href="blah">aba<table><tr><td><a href="foo">br</td></tr>x</table>aoe
Got:
| <html>
|   <head>
|   <body>
|     <a>
|       href="blah"
|       "aba"
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <a>
|                 href="foo"
|                 "br"
|           "x"
|       "aoe"
Expected:
| <html>
|   <head>
|   <body>
|     <a>
|       href="blah"
|       "abax"
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <a>
|                 href="foo"
|                 "br"
|       "aoe"

Test 80 of 113 in resources/tests1.dat failed. Input:
<table><a href="blah">aba<tr><td><a href="foo">br</td></tr>x</table>aoe
Got:
| <html>
|   <head>
|   <body>
|     <a>
|       href="blah"
|       "aba"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <a>
|               href="foo"
|               "br"
|         "x"
|     "aoe"
Expected:
| <html>
|   <head>
|   <body>
|     <a>
|       href="blah"
|       "aba"
|     <a>
|       href="blah"
|       "x"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <a>
|               href="foo"
|               "br"
|     <a>
|       href="blah"
|       "aoe"

Test 103 of 113 in resources/tests1.dat failed. Input:
<a><table><td><a><table></table><a></tr><a></table><a>
Got:
| <html>
|   <head>
|   <body>
|     <a>
|       <a>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <a>
|                 <table>
|               <a>
|       <a>
Expected:
| <html>
|   <head>
|   <body>
|     <a>
|       <a>
|       <table>
|         <tbody>
|           <tr>
|             <td>
|               <a>
|                 <table>
|               <a>
|     <a>
resources/tests2.dat:
13

Test 13 of 59 in resources/tests2.dat failed. Input:
<!DOCTYPE html><table><tr>TEST
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           "TEST"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "TEST"
|     <table>
|       <tbody>
|         <tr>
resources/tests3.dat:
16
17
18

Test 16 of 23 in resources/tests3.dat failed. Input:
<!DOCTYPE html><textarea>
</textarea>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "
"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>

Test 17 of 23 in resources/tests3.dat failed. Input:
<!DOCTYPE html><textarea>
foo</textarea>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "
foo"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "foo"

Test 18 of 23 in resources/tests3.dat failed. Input:
<!DOCTYPE html><textarea>

foo</textarea>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "

foo"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "
foo"
resources/tests4.dat:
3
5
6
7

Test 3 of 7 in resources/tests4.dat failed. Input:
textarea content with <em>pseudo</em> <foo>markup
Got:
| "textarea content with "
| <em>
|   "pseudo"
| " "
| <foo>
|   "markup"
Expected:
| "textarea content with <em>pseudo</em> <foo>markup"

Test 5 of 7 in resources/tests4.dat failed. Input:
</plaintext>
Got:
| 
Expected:
| "</plaintext>"

Test 6 of 7 in resources/tests4.dat failed. Input:
setting html's innerHTML
Got:
| "setting html's innerHTML"
Expected:
| <head>
| <body>
|   "setting html's innerHTML"

Test 7 of 7 in resources/tests4.dat failed. Input:
<title>setting head's innerHTML</title>
Got:
| 
Expected:
| <title>
|   "setting head's innerHTML"
resources/tests5.dat: PASS

resources/tests6.dat:
26
27
30
37
44
45
47

Test 26 of 51 in resources/tests6.dat failed. Input:
<table><colgroup>foo
Got:
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>
|       "foo"
Expected:
| <html>
|   <head>
|   <body>
|     "foo"
|     <table>
|       <colgroup>

Test 27 of 51 in resources/tests6.dat failed. Input:
foo<col>
Got:
| 
Expected:
| <col>

Test 30 of 51 in resources/tests6.dat failed. Input:
</frameset><frame>
Got:
| 
Expected:
| <frame>

Test 37 of 51 in resources/tests6.dat failed. Input:
<caption><col><colgroup><tbody><tfoot><thead><tr>
Got:
| <caption>
| <colgroup>
| <tbody>
| <tfoot>
| <thead>
|   <tr>
Expected:
| <tr>

Test 44 of 51 in resources/tests6.dat failed. Input:
</table><tr>
Got:
| <tr>
Expected:
| <tbody>
|   <tr>

Test 45 of 51 in resources/tests6.dat failed. Input:
<body></body></html>
Got:
| 
Expected:
| <head>
| <body>

Test 47 of 51 in resources/tests6.dat failed. Input:
<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN"><html></html>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
Expected:
| <!DOCTYPE html "-//W3C//DTD HTML 4.01//EN" "">
| <html>
|   <head>
|   <body>
resources/tests7.dat:
6
16
24
27
28
29
30

Test 6 of 30 in resources/tests7.dat failed. Input:
<!doctype html><table>X<tr><td><table> <meta></table></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       "X"
|       <tbody>
|         <tr>
|           <td>
|             <meta>
|             <table>
|               " "
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "X"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <meta>
|             <table>
|               " "

Test 16 of 30 in resources/tests7.dat failed. Input:
<!doctype html><table>X<input type=hidDEN></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       "X"
|       <input>
|         type="hidDEN"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "X"
|     <table>
|       <input>
|         type="hidDEN"

Test 24 of 30 in resources/tests7.dat failed. Input:
<body>X</body></body>
Got:
| "X"
Expected:
| <head>
| <body>
|   "X"

Test 27 of 30 in resources/tests7.dat failed. Input:
<table><b><tr><td>aaa</td></tr>bbb</table>ccc
Got:
| <html>
|   <head>
|   <body>
|     <b>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "aaa"
|         "bbb"
|     "ccc"
Expected:
| <html>
|   <head>
|   <body>
|     <b>
|     <b>
|       "bbb"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "aaa"
|     <b>
|       "ccc"

Test 28 of 30 in resources/tests7.dat failed. Input:
A<table><tr> B</tr> B</table>
Got:
| <html>
|   <head>
|   <body>
|     "A"
|     <table>
|       <tbody>
|         <tr>
|           " B"
|         " B"
Expected:
| <html>
|   <head>
|   <body>
|     "A B B"
|     <table>
|       <tbody>
|         <tr>

Test 29 of 30 in resources/tests7.dat failed. Input:
A<table><tr> B</tr> </em>C</table>
Got:
| <html>
|   <head>
|   <body>
|     "A"
|     <table>
|       <tbody>
|         <tr>
|           " B"
|         " C"
Expected:
| <html>
|   <head>
|   <body>
|     "A BC"
|     <table>
|       <tbody>
|         <tr>
|         " "

Test 30 of 30 in resources/tests7.dat failed. Input:
<select><keygen>
Got:
| <html>
|   <head>
|   <body>
|     <select>
|     <keygen>
|       <option>
|         "2048 (High Grade)"
|       <option>
|         "1024 (Medium Grade)"
|       <option>
|         "512 (Low Grade)"
Expected:
| <html>
|   <head>
|   <body>
|     <select>
|     <keygen>
resources/tests8.dat:
6
7

Test 6 of 9 in resources/tests8.dat failed. Input:
x<table>x
Got:
| <html>
|   <head>
|   <body>
|     "x"
|     <table>
|       "x"
Expected:
| <html>
|   <head>
|   <body>
|     "xx"
|     <table>

Test 7 of 9 in resources/tests8.dat failed. Input:
x<table><table>x
Got:
| <html>
|   <head>
|   <body>
|     "x"
|     <table>
|     <table>
|       "x"
Expected:
| <html>
|   <head>
|   <body>
|     "x"
|     <table>
|     "x"
|     <table>
resources/tests9.dat:
1
2
5
6
7
8
9
10
11
12
13
14
15
18
19
22
23
24
25

Test 1 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><math></math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>

Test 2 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><math></math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>

Test 5 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><math></math></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|     <table>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|     <table>

Test 6 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><math><mi>foo</mi></math></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>
|         "foo"
|     <table>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|     <table>

Test 7 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><math><mi>foo</mi><mi>bar</mi></math></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>
|         "foo"
|       <mi>
|         "bar"
|     <table>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <table>

Test 8 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><tbody><math><mi>foo</mi><mi>bar</mi></math></tbody></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>
|         "foo"
|       <mi>
|         "bar"
|     <table>
|       <tbody>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <table>
|       <tbody>

Test 9 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><tbody><tr><math><mi>foo</mi><mi>bar</mi></math></tr></tbody></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>
|         "foo"
|       <mi>
|         "bar"
|     <table>
|       <tbody>
|         <tr>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <table>
|       <tbody>
|         <tr>

Test 10 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><tbody><tr><td><math><mi>foo</mi><mi>bar</mi></math></td></tr></tbody></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <math>
|               <mi>
|                 "foo"
|               <mi>
|                 "bar"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <math math>
|               <math mi>
|                 "foo"
|               <math mi>
|                 "bar"

Test 11 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><tbody><tr><td><math><mi>foo</mi><mi>bar</mi></math><p>baz</td></tr></tbody></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <math>
|               <mi>
|                 "foo"
|               <mi>
|                 "bar"
|             <p>
|               "baz"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <math math>
|               <math mi>
|                 "foo"
|               <math mi>
|                 "bar"
|             <p>
|               "baz"

Test 12 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><caption><math><mi>foo</mi><mi>bar</mi></math><p>baz</caption></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math>
|           <mi>
|             "foo"
|           <mi>
|             "bar"
|         <p>
|           "baz"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math math>
|           <math mi>
|             "foo"
|           <math mi>
|             "bar"
|         <p>
|           "baz"

Test 13 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><caption><math><mi>foo</mi><mi>bar</mi><p>baz</table><p>quux
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math>
|           <mi>
|             "foo"
|           <mi>
|             "bar"
|           <p>
|             "baz"
|     <p>
|       "quux"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math math>
|           <math mi>
|             "foo"
|           <math mi>
|             "bar"
|         <p>
|           "baz"
|     <p>
|       "quux"

Test 14 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><caption><math><mi>foo</mi><mi>bar</mi>baz</table><p>quux
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math>
|           <mi>
|             "foo"
|           <mi>
|             "bar"
|           "baz"
|     <p>
|       "quux"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <math math>
|           <math mi>
|             "foo"
|           <math mi>
|             "bar"
|           "baz"
|     <p>
|       "quux"

Test 15 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body><table><colgroup><math><mi>foo</mi><mi>bar</mi><p>baz</table><p>quux
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>
|         "foo"
|       <mi>
|         "bar"
|       <p>
|         "baz"
|     <table>
|       <colgroup>
|     <p>
|       "quux"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <p>
|       "baz"
|     <table>
|       <colgroup>
|     <p>
|       "quux"

Test 18 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body></body></html><math><mi>foo</mi><mi>bar</mi><p>baz
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>
|         "foo"
|       <mi>
|         "bar"
|       <p>
|         "baz"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <p>
|       "baz"

Test 19 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body></body><math><mi>foo</mi><mi>bar</mi><p>baz
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <mi>
|         "foo"
|       <mi>
|         "bar"
|       <p>
|         "baz"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "foo"
|       <math mi>
|         "bar"
|     <p>
|       "baz"

Test 22 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body xlink:href=foo><math xlink:href=foo></math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     <math>
|       xlink:href="foo"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     <math math>
|       xlink href="foo"

Test 23 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body xlink:href=foo xml:lang=en><math><mi xml:lang=en xlink:href=foo></mi></math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math>
|       <mi>
|         xlink:href="foo"
|         xml:lang="en"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math math>
|       <math mi>
|         xlink href="foo"
|         xml lang="en"

Test 24 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body xlink:href=foo xml:lang=en><math><mi xml:lang=en xlink:href=foo /></math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math>
|       <mi>
|         xlink:href="foo"
|         xml:lang="en"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math math>
|       <math mi>
|         xlink href="foo"
|         xml lang="en"

Test 25 of 25 in resources/tests9.dat failed. Input:
<!DOCTYPE html><body xlink:href=foo xml:lang=en><math><mi xml:lang=en xlink:href=foo />bar</math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math>
|       <mi>
|         xlink:href="foo"
|         xml:lang="en"
|         "bar"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <math math>
|       <math mi>
|         xlink href="foo"
|         xml lang="en"
|       "bar"
resources/tests10.dat:
1
2
5
6
7
8
9
10
11
12
13
14
15
18
19
22
23
24
25

Test 1 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><svg></svg>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>

Test 2 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><svg></svg>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>

Test 5 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><svg></svg></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|     <table>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <table>

Test 6 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><svg><g>foo</g></svg></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|       <g>
|         "foo"
|     <table>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|     <table>

Test 7 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><svg><g>foo</g><g>bar</g></svg></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|       <g>
|         "foo"
|       <g>
|         "bar"
|     <table>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <table>

Test 8 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><tbody><svg><g>foo</g><g>bar</g></svg></tbody></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|       <g>
|         "foo"
|       <g>
|         "bar"
|     <table>
|       <tbody>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <table>
|       <tbody>

Test 9 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><tbody><tr><svg><g>foo</g><g>bar</g></svg></tr></tbody></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|       <g>
|         "foo"
|       <g>
|         "bar"
|     <table>
|       <tbody>
|         <tr>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <table>
|       <tbody>
|         <tr>

Test 10 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><tbody><tr><td><svg><g>foo</g><g>bar</g></svg></td></tr></tbody></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg>
|               <g>
|                 "foo"
|               <g>
|                 "bar"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "foo"
|               <svg g>
|                 "bar"

Test 11 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><tbody><tr><td><svg><g>foo</g><g>bar</g></svg><p>baz</td></tr></tbody></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg>
|               <g>
|                 "foo"
|               <g>
|                 "bar"
|             <p>
|               "baz"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <svg svg>
|               <svg g>
|                 "foo"
|               <svg g>
|                 "bar"
|             <p>
|               "baz"

Test 12 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><caption><svg><g>foo</g><g>bar</g></svg><p>baz</caption></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg>
|           <g>
|             "foo"
|           <g>
|             "bar"
|         <p>
|           "baz"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "foo"
|           <svg g>
|             "bar"
|         <p>
|           "baz"

Test 13 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><caption><svg><g>foo</g><g>bar</g><p>baz</table><p>quux
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg>
|           <g>
|             "foo"
|           <g>
|             "bar"
|           <p>
|             "baz"
|     <p>
|       "quux"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "foo"
|           <svg g>
|             "bar"
|         <p>
|           "baz"
|     <p>
|       "quux"

Test 14 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><caption><svg><g>foo</g><g>bar</g>baz</table><p>quux
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg>
|           <g>
|             "foo"
|           <g>
|             "bar"
|           "baz"
|     <p>
|       "quux"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <svg svg>
|           <svg g>
|             "foo"
|           <svg g>
|             "bar"
|           "baz"
|     <p>
|       "quux"

Test 15 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><colgroup><svg><g>foo</g><g>bar</g><p>baz</table><p>quux
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|       <g>
|         "foo"
|       <g>
|         "bar"
|       <p>
|         "baz"
|     <table>
|       <colgroup>
|     <p>
|       "quux"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <p>
|       "baz"
|     <table>
|       <colgroup>
|     <p>
|       "quux"

Test 18 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body></body></html><svg><g>foo</g><g>bar</g><p>baz
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|       <g>
|         "foo"
|       <g>
|         "bar"
|       <p>
|         "baz"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <p>
|       "baz"

Test 19 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body></body><svg><g>foo</g><g>bar</g><p>baz
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|       <g>
|         "foo"
|       <g>
|         "bar"
|       <p>
|         "baz"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         "foo"
|       <svg g>
|         "bar"
|     <p>
|       "baz"

Test 22 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body xlink:href=foo><svg xlink:href=foo></svg>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     <svg>
|       xlink:href="foo"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     <svg svg>
|       xlink href="foo"

Test 23 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body xlink:href=foo xml:lang=en><svg><g xml:lang=en xlink:href=foo></g></svg>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <svg>
|       <g>
|         xlink:href="foo"
|         xml:lang="en"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <svg svg>
|       <svg g>
|         xlink href="foo"
|         xml lang="en"

Test 24 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body xlink:href=foo xml:lang=en><svg><g xml:lang=en xlink:href=foo /></svg>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <svg>
|       <g>
|         xlink:href="foo"
|         xml:lang="en"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <svg svg>
|       <svg g>
|         xlink href="foo"
|         xml lang="en"

Test 25 of 25 in resources/tests10.dat failed. Input:
<!DOCTYPE html><body xlink:href=foo xml:lang=en><svg><g xml:lang=en xlink:href=foo />bar</svg>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <svg>
|       <g>
|         xlink:href="foo"
|         xml:lang="en"
|         "bar"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     xlink:href="foo"
|     xml:lang="en"
|     <svg svg>
|       <svg g>
|         xlink href="foo"
|         xml lang="en"
|       "bar"
resources/tests11.dat:
1
2
3
4
5
6
7
8
9

Test 1 of 9 in resources/tests11.dat failed. Input:
<!DOCTYPE html><body><svg attributeName='' attributeType='' baseFrequency='' baseProfile='' calcMode='' clipPathUnits='' contentScriptType='' contentStyleType='' diffuseConstant='' edgeMode='' externalResourcesRequired='' filterRes='' filterUnits='' glyphRef='' gradientTransform='' gradientUnits='' kernelMatrix='' kernelUnitLength='' keyPoints='' keySplines='' keyTimes='' lengthAdjust='' limitingConeAngle='' markerHeight='' markerUnits='' markerWidth='' maskContentUnits='' maskUnits='' numOctaves='' pathLength='' patternContentUnits='' patternTransform='' patternUnits='' pointsAtX='' pointsAtY='' pointsAtZ='' preserveAlpha='' preserveAspectRatio='' primitiveUnits='' refX='' refY='' repeatCount='' repeatDur='' requiredExtensions='' requiredFeatures='' specularConstant='' specularExponent='' spreadMethod='' startOffset='' stdDeviation='' stitchTiles='' surfaceScale='' systemLanguage='' tableValues='' targetX='' targetY='' textLength='' viewBox='' viewTarget='' xChannelSelector='' yChannelSelector='' zoomAndPan=''></svg>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|       attributename=""
|       attributetype=""
|       basefrequency=""
|       baseprofile=""
|       calcmode=""
|       clippathunits=""
|       contentscripttype=""
|       contentstyletype=""
|       diffuseconstant=""
|       edgemode=""
|       externalresourcesrequired=""
|       filterres=""
|       filterunits=""
|       glyphref=""
|       gradienttransform=""
|       gradientunits=""
|       kernelmatrix=""
|       kernelunitlength=""
|       keypoints=""
|       keysplines=""
|       keytimes=""
|       lengthadjust=""
|       limitingconeangle=""
|       markerheight=""
|       markerunits=""
|       markerwidth=""
|       maskcontentunits=""
|       maskunits=""
|       numoctaves=""
|       pathlength=""
|       patterncontentunits=""
|       patterntransform=""
|       patternunits=""
|       pointsatx=""
|       pointsaty=""
|       pointsatz=""
|       preservealpha=""
|       preserveaspectratio=""
|       primitiveunits=""
|       refx=""
|       refy=""
|       repeatcount=""
|       repeatdur=""
|       requiredextensions=""
|       requiredfeatures=""
|       specularconstant=""
|       specularexponent=""
|       spreadmethod=""
|       startoffset=""
|       stddeviation=""
|       stitchtiles=""
|       surfacescale=""
|       systemlanguage=""
|       tablevalues=""
|       targetx=""
|       targety=""
|       textlength=""
|       viewbox=""
|       viewtarget=""
|       xchannelselector=""
|       ychannelselector=""
|       zoomandpan=""
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       attributeName=""
|       attributeType=""
|       baseFrequency=""
|       baseProfile=""
|       calcMode=""
|       clipPathUnits=""
|       contentScriptType=""
|       contentStyleType=""
|       diffuseConstant=""
|       edgeMode=""
|       externalResourcesRequired=""
|       filterRes=""
|       filterUnits=""
|       glyphRef=""
|       gradientTransform=""
|       gradientUnits=""
|       kernelMatrix=""
|       kernelUnitLength=""
|       keyPoints=""
|       keySplines=""
|       keyTimes=""
|       lengthAdjust=""
|       limitingConeAngle=""
|       markerHeight=""
|       markerUnits=""
|       markerWidth=""
|       maskContentUnits=""
|       maskUnits=""
|       numOctaves=""
|       pathLength=""
|       patternContentUnits=""
|       patternTransform=""
|       patternUnits=""
|       pointsAtX=""
|       pointsAtY=""
|       pointsAtZ=""
|       preserveAlpha=""
|       preserveAspectRatio=""
|       primitiveUnits=""
|       refX=""
|       refY=""
|       repeatCount=""
|       repeatDur=""
|       requiredExtensions=""
|       requiredFeatures=""
|       specularConstant=""
|       specularExponent=""
|       spreadMethod=""
|       startOffset=""
|       stdDeviation=""
|       stitchTiles=""
|       surfaceScale=""
|       systemLanguage=""
|       tableValues=""
|       targetX=""
|       targetY=""
|       textLength=""
|       viewBox=""
|       viewTarget=""
|       xChannelSelector=""
|       yChannelSelector=""
|       zoomAndPan=""

Test 2 of 9 in resources/tests11.dat failed. Input:
<!DOCTYPE html><BODY><SVG ATTRIBUTENAME='' ATTRIBUTETYPE='' BASEFREQUENCY='' BASEPROFILE='' CALCMODE='' CLIPPATHUNITS='' CONTENTSCRIPTTYPE='' CONTENTSTYLETYPE='' DIFFUSECONSTANT='' EDGEMODE='' EXTERNALRESOURCESREQUIRED='' FILTERRES='' FILTERUNITS='' GLYPHREF='' GRADIENTTRANSFORM='' GRADIENTUNITS='' KERNELMATRIX='' KERNELUNITLENGTH='' KEYPOINTS='' KEYSPLINES='' KEYTIMES='' LENGTHADJUST='' LIMITINGCONEANGLE='' MARKERHEIGHT='' MARKERUNITS='' MARKERWIDTH='' MASKCONTENTUNITS='' MASKUNITS='' NUMOCTAVES='' PATHLENGTH='' PATTERNCONTENTUNITS='' PATTERNTRANSFORM='' PATTERNUNITS='' POINTSATX='' POINTSATY='' POINTSATZ='' PRESERVEALPHA='' PRESERVEASPECTRATIO='' PRIMITIVEUNITS='' REFX='' REFY='' REPEATCOUNT='' REPEATDUR='' REQUIREDEXTENSIONS='' REQUIREDFEATURES='' SPECULARCONSTANT='' SPECULAREXPONENT='' SPREADMETHOD='' STARTOFFSET='' STDDEVIATION='' STITCHTILES='' SURFACESCALE='' SYSTEMLANGUAGE='' TABLEVALUES='' TARGETX='' TARGETY='' TEXTLENGTH='' VIEWBOX='' VIEWTARGET='' XCHANNELSELECTOR='' YCHANNELSELECTOR='' ZOOMANDPAN=''></SVG>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|       attributename=""
|       attributetype=""
|       basefrequency=""
|       baseprofile=""
|       calcmode=""
|       clippathunits=""
|       contentscripttype=""
|       contentstyletype=""
|       diffuseconstant=""
|       edgemode=""
|       externalresourcesrequired=""
|       filterres=""
|       filterunits=""
|       glyphref=""
|       gradienttransform=""
|       gradientunits=""
|       kernelmatrix=""
|       kernelunitlength=""
|       keypoints=""
|       keysplines=""
|       keytimes=""
|       lengthadjust=""
|       limitingconeangle=""
|       markerheight=""
|       markerunits=""
|       markerwidth=""
|       maskcontentunits=""
|       maskunits=""
|       numoctaves=""
|       pathlength=""
|       patterncontentunits=""
|       patterntransform=""
|       patternunits=""
|       pointsatx=""
|       pointsaty=""
|       pointsatz=""
|       preservealpha=""
|       preserveaspectratio=""
|       primitiveunits=""
|       refx=""
|       refy=""
|       repeatcount=""
|       repeatdur=""
|       requiredextensions=""
|       requiredfeatures=""
|       specularconstant=""
|       specularexponent=""
|       spreadmethod=""
|       startoffset=""
|       stddeviation=""
|       stitchtiles=""
|       surfacescale=""
|       systemlanguage=""
|       tablevalues=""
|       targetx=""
|       targety=""
|       textlength=""
|       viewbox=""
|       viewtarget=""
|       xchannelselector=""
|       ychannelselector=""
|       zoomandpan=""
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       attributeName=""
|       attributeType=""
|       baseFrequency=""
|       baseProfile=""
|       calcMode=""
|       clipPathUnits=""
|       contentScriptType=""
|       contentStyleType=""
|       diffuseConstant=""
|       edgeMode=""
|       externalResourcesRequired=""
|       filterRes=""
|       filterUnits=""
|       glyphRef=""
|       gradientTransform=""
|       gradientUnits=""
|       kernelMatrix=""
|       kernelUnitLength=""
|       keyPoints=""
|       keySplines=""
|       keyTimes=""
|       lengthAdjust=""
|       limitingConeAngle=""
|       markerHeight=""
|       markerUnits=""
|       markerWidth=""
|       maskContentUnits=""
|       maskUnits=""
|       numOctaves=""
|       pathLength=""
|       patternContentUnits=""
|       patternTransform=""
|       patternUnits=""
|       pointsAtX=""
|       pointsAtY=""
|       pointsAtZ=""
|       preserveAlpha=""
|       preserveAspectRatio=""
|       primitiveUnits=""
|       refX=""
|       refY=""
|       repeatCount=""
|       repeatDur=""
|       requiredExtensions=""
|       requiredFeatures=""
|       specularConstant=""
|       specularExponent=""
|       spreadMethod=""
|       startOffset=""
|       stdDeviation=""
|       stitchTiles=""
|       surfaceScale=""
|       systemLanguage=""
|       tableValues=""
|       targetX=""
|       targetY=""
|       textLength=""
|       viewBox=""
|       viewTarget=""
|       xChannelSelector=""
|       yChannelSelector=""
|       zoomAndPan=""

Test 3 of 9 in resources/tests11.dat failed. Input:
<!DOCTYPE html><body><svg attributename='' attributetype='' basefrequency='' baseprofile='' calcmode='' clippathunits='' contentscripttype='' contentstyletype='' diffuseconstant='' edgemode='' externalresourcesrequired='' filterres='' filterunits='' glyphref='' gradienttransform='' gradientunits='' kernelmatrix='' kernelunitlength='' keypoints='' keysplines='' keytimes='' lengthadjust='' limitingconeangle='' markerheight='' markerunits='' markerwidth='' maskcontentunits='' maskunits='' numoctaves='' pathlength='' patterncontentunits='' patterntransform='' patternunits='' pointsatx='' pointsaty='' pointsatz='' preservealpha='' preserveaspectratio='' primitiveunits='' refx='' refy='' repeatcount='' repeatdur='' requiredextensions='' requiredfeatures='' specularconstant='' specularexponent='' spreadmethod='' startoffset='' stddeviation='' stitchtiles='' surfacescale='' systemlanguage='' tablevalues='' targetx='' targety='' textlength='' viewbox='' viewtarget='' xchannelselector='' ychannelselector='' zoomandpan=''></svg>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|       attributename=""
|       attributetype=""
|       basefrequency=""
|       baseprofile=""
|       calcmode=""
|       clippathunits=""
|       contentscripttype=""
|       contentstyletype=""
|       diffuseconstant=""
|       edgemode=""
|       externalresourcesrequired=""
|       filterres=""
|       filterunits=""
|       glyphref=""
|       gradienttransform=""
|       gradientunits=""
|       kernelmatrix=""
|       kernelunitlength=""
|       keypoints=""
|       keysplines=""
|       keytimes=""
|       lengthadjust=""
|       limitingconeangle=""
|       markerheight=""
|       markerunits=""
|       markerwidth=""
|       maskcontentunits=""
|       maskunits=""
|       numoctaves=""
|       pathlength=""
|       patterncontentunits=""
|       patterntransform=""
|       patternunits=""
|       pointsatx=""
|       pointsaty=""
|       pointsatz=""
|       preservealpha=""
|       preserveaspectratio=""
|       primitiveunits=""
|       refx=""
|       refy=""
|       repeatcount=""
|       repeatdur=""
|       requiredextensions=""
|       requiredfeatures=""
|       specularconstant=""
|       specularexponent=""
|       spreadmethod=""
|       startoffset=""
|       stddeviation=""
|       stitchtiles=""
|       surfacescale=""
|       systemlanguage=""
|       tablevalues=""
|       targetx=""
|       targety=""
|       textlength=""
|       viewbox=""
|       viewtarget=""
|       xchannelselector=""
|       ychannelselector=""
|       zoomandpan=""
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       attributeName=""
|       attributeType=""
|       baseFrequency=""
|       baseProfile=""
|       calcMode=""
|       clipPathUnits=""
|       contentScriptType=""
|       contentStyleType=""
|       diffuseConstant=""
|       edgeMode=""
|       externalResourcesRequired=""
|       filterRes=""
|       filterUnits=""
|       glyphRef=""
|       gradientTransform=""
|       gradientUnits=""
|       kernelMatrix=""
|       kernelUnitLength=""
|       keyPoints=""
|       keySplines=""
|       keyTimes=""
|       lengthAdjust=""
|       limitingConeAngle=""
|       markerHeight=""
|       markerUnits=""
|       markerWidth=""
|       maskContentUnits=""
|       maskUnits=""
|       numOctaves=""
|       pathLength=""
|       patternContentUnits=""
|       patternTransform=""
|       patternUnits=""
|       pointsAtX=""
|       pointsAtY=""
|       pointsAtZ=""
|       preserveAlpha=""
|       preserveAspectRatio=""
|       primitiveUnits=""
|       refX=""
|       refY=""
|       repeatCount=""
|       repeatDur=""
|       requiredExtensions=""
|       requiredFeatures=""
|       specularConstant=""
|       specularExponent=""
|       spreadMethod=""
|       startOffset=""
|       stdDeviation=""
|       stitchTiles=""
|       surfaceScale=""
|       systemLanguage=""
|       tableValues=""
|       targetX=""
|       targetY=""
|       textLength=""
|       viewBox=""
|       viewTarget=""
|       xChannelSelector=""
|       yChannelSelector=""
|       zoomAndPan=""

Test 4 of 9 in resources/tests11.dat failed. Input:
<!DOCTYPE html><body><math attributeName='' attributeType='' baseFrequency='' baseProfile='' calcMode='' clipPathUnits='' contentScriptType='' contentStyleType='' diffuseConstant='' edgeMode='' externalResourcesRequired='' filterRes='' filterUnits='' glyphRef='' gradientTransform='' gradientUnits='' kernelMatrix='' kernelUnitLength='' keyPoints='' keySplines='' keyTimes='' lengthAdjust='' limitingConeAngle='' markerHeight='' markerUnits='' markerWidth='' maskContentUnits='' maskUnits='' numOctaves='' pathLength='' patternContentUnits='' patternTransform='' patternUnits='' pointsAtX='' pointsAtY='' pointsAtZ='' preserveAlpha='' preserveAspectRatio='' primitiveUnits='' refX='' refY='' repeatCount='' repeatDur='' requiredExtensions='' requiredFeatures='' specularConstant='' specularExponent='' spreadMethod='' startOffset='' stdDeviation='' stitchTiles='' surfaceScale='' systemLanguage='' tableValues='' targetX='' targetY='' textLength='' viewBox='' viewTarget='' xChannelSelector='' yChannelSelector='' zoomAndPan=''></math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       attributename=""
|       attributetype=""
|       basefrequency=""
|       baseprofile=""
|       calcmode=""
|       clippathunits=""
|       contentscripttype=""
|       contentstyletype=""
|       diffuseconstant=""
|       edgemode=""
|       externalresourcesrequired=""
|       filterres=""
|       filterunits=""
|       glyphref=""
|       gradienttransform=""
|       gradientunits=""
|       kernelmatrix=""
|       kernelunitlength=""
|       keypoints=""
|       keysplines=""
|       keytimes=""
|       lengthadjust=""
|       limitingconeangle=""
|       markerheight=""
|       markerunits=""
|       markerwidth=""
|       maskcontentunits=""
|       maskunits=""
|       numoctaves=""
|       pathlength=""
|       patterncontentunits=""
|       patterntransform=""
|       patternunits=""
|       pointsatx=""
|       pointsaty=""
|       pointsatz=""
|       preservealpha=""
|       preserveaspectratio=""
|       primitiveunits=""
|       refx=""
|       refy=""
|       repeatcount=""
|       repeatdur=""
|       requiredextensions=""
|       requiredfeatures=""
|       specularconstant=""
|       specularexponent=""
|       spreadmethod=""
|       startoffset=""
|       stddeviation=""
|       stitchtiles=""
|       surfacescale=""
|       systemlanguage=""
|       tablevalues=""
|       targetx=""
|       targety=""
|       textlength=""
|       viewbox=""
|       viewtarget=""
|       xchannelselector=""
|       ychannelselector=""
|       zoomandpan=""
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       attributename=""
|       attributetype=""
|       basefrequency=""
|       baseprofile=""
|       calcmode=""
|       clippathunits=""
|       contentscripttype=""
|       contentstyletype=""
|       diffuseconstant=""
|       edgemode=""
|       externalresourcesrequired=""
|       filterres=""
|       filterunits=""
|       glyphref=""
|       gradienttransform=""
|       gradientunits=""
|       kernelmatrix=""
|       kernelunitlength=""
|       keypoints=""
|       keysplines=""
|       keytimes=""
|       lengthadjust=""
|       limitingconeangle=""
|       markerheight=""
|       markerunits=""
|       markerwidth=""
|       maskcontentunits=""
|       maskunits=""
|       numoctaves=""
|       pathlength=""
|       patterncontentunits=""
|       patterntransform=""
|       patternunits=""
|       pointsatx=""
|       pointsaty=""
|       pointsatz=""
|       preservealpha=""
|       preserveaspectratio=""
|       primitiveunits=""
|       refx=""
|       refy=""
|       repeatcount=""
|       repeatdur=""
|       requiredextensions=""
|       requiredfeatures=""
|       specularconstant=""
|       specularexponent=""
|       spreadmethod=""
|       startoffset=""
|       stddeviation=""
|       stitchtiles=""
|       surfacescale=""
|       systemlanguage=""
|       tablevalues=""
|       targetx=""
|       targety=""
|       textlength=""
|       viewbox=""
|       viewtarget=""
|       xchannelselector=""
|       ychannelselector=""
|       zoomandpan=""

Test 5 of 9 in resources/tests11.dat failed. Input:
<!DOCTYPE html><body><svg><altGlyph /><altGlyphDef /><altGlyphItem /><animateColor /><animateMotion /><animateTransform /><clipPath /><feBlend /><feColorMatrix /><feComponentTransfer /><feComposite /><feConvolveMatrix /><feDiffuseLighting /><feDisplacementMap /><feDistantLight /><feFlood /><feFuncA /><feFuncB /><feFuncG /><feFuncR /><feGaussianBlur /><feImage /><feMerge /><feMergeNode /><feMorphology /><feOffset /><fePointLight /><feSpecularLighting /><feSpotLight /><feTile /><feTurbulence /><foreignObject /><glyphRef /><linearGradient /><radialGradient /><textPath /></svg>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|       <altglyph>
|         <altglyphdef>
|           <altglyphitem>
|             <animatecolor>
|               <animatemotion>
|                 <animatetransform>
|                   <clippath>
|                     <feblend>
|                       <fecolormatrix>
|                         <fecomponenttransfer>
|                           <fecomposite>
|                             <feconvolvematrix>
|                               <fediffuselighting>
|                                 <fedisplacementmap>
|                                   <fedistantlight>
|                                     <feflood>
|                                       <fefunca>
|                                         <fefuncb>
|                                           <fefuncg>
|                                             <fefuncr>
|                                               <fegaussianblur>
|                                                 <feimage>
|                                                   <femerge>
|                                                     <femergenode>
|                                                       <femorphology>
|                                                         <feoffset>
|                                                           <fepointlight>
|                                                             <fespecularlighting>
|                                                               <fespotlight>
|                                                                 <fetile>
|                                                                   <feturbulence>
|                                                                     <foreignobject>
|                                                                       <glyphref>
|                                                                         <lineargradient>
|                                                                           <radialgradient>
|                                                                             <textpath>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg altGlyph>
|       <svg altGlyphDef>
|       <svg altGlyphItem>
|       <svg animateColor>
|       <svg animateMotion>
|       <svg animateTransform>
|       <svg clipPath>
|       <svg feBlend>
|       <svg feColorMatrix>
|       <svg feComponentTransfer>
|       <svg feComposite>
|       <svg feConvolveMatrix>
|       <svg feDiffuseLighting>
|       <svg feDisplacementMap>
|       <svg feDistantLight>
|       <svg feFlood>
|       <svg feFuncA>
|       <svg feFuncB>
|       <svg feFuncG>
|       <svg feFuncR>
|       <svg feGaussianBlur>
|       <svg feImage>
|       <svg feMerge>
|       <svg feMergeNode>
|       <svg feMorphology>
|       <svg feOffset>
|       <svg fePointLight>
|       <svg feSpecularLighting>
|       <svg feSpotLight>
|       <svg feTile>
|       <svg feTurbulence>
|       <svg foreignObject>
|       <svg glyphRef>
|       <svg linearGradient>
|       <svg radialGradient>
|       <svg textPath>

Test 6 of 9 in resources/tests11.dat failed. Input:
<!DOCTYPE html><body><svg><altglyph /><altglyphdef /><altglyphitem /><animatecolor /><animatemotion /><animatetransform /><clippath /><feblend /><fecolormatrix /><fecomponenttransfer /><fecomposite /><feconvolvematrix /><fediffuselighting /><fedisplacementmap /><fedistantlight /><feflood /><fefunca /><fefuncb /><fefuncg /><fefuncr /><fegaussianblur /><feimage /><femerge /><femergenode /><femorphology /><feoffset /><fepointlight /><fespecularlighting /><fespotlight /><fetile /><feturbulence /><foreignobject /><glyphref /><lineargradient /><radialgradient /><textpath /></svg>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|       <altglyph>
|         <altglyphdef>
|           <altglyphitem>
|             <animatecolor>
|               <animatemotion>
|                 <animatetransform>
|                   <clippath>
|                     <feblend>
|                       <fecolormatrix>
|                         <fecomponenttransfer>
|                           <fecomposite>
|                             <feconvolvematrix>
|                               <fediffuselighting>
|                                 <fedisplacementmap>
|                                   <fedistantlight>
|                                     <feflood>
|                                       <fefunca>
|                                         <fefuncb>
|                                           <fefuncg>
|                                             <fefuncr>
|                                               <fegaussianblur>
|                                                 <feimage>
|                                                   <femerge>
|                                                     <femergenode>
|                                                       <femorphology>
|                                                         <feoffset>
|                                                           <fepointlight>
|                                                             <fespecularlighting>
|                                                               <fespotlight>
|                                                                 <fetile>
|                                                                   <feturbulence>
|                                                                     <foreignobject>
|                                                                       <glyphref>
|                                                                         <lineargradient>
|                                                                           <radialgradient>
|                                                                             <textpath>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg altGlyph>
|       <svg altGlyphDef>
|       <svg altGlyphItem>
|       <svg animateColor>
|       <svg animateMotion>
|       <svg animateTransform>
|       <svg clipPath>
|       <svg feBlend>
|       <svg feColorMatrix>
|       <svg feComponentTransfer>
|       <svg feComposite>
|       <svg feConvolveMatrix>
|       <svg feDiffuseLighting>
|       <svg feDisplacementMap>
|       <svg feDistantLight>
|       <svg feFlood>
|       <svg feFuncA>
|       <svg feFuncB>
|       <svg feFuncG>
|       <svg feFuncR>
|       <svg feGaussianBlur>
|       <svg feImage>
|       <svg feMerge>
|       <svg feMergeNode>
|       <svg feMorphology>
|       <svg feOffset>
|       <svg fePointLight>
|       <svg feSpecularLighting>
|       <svg feSpotLight>
|       <svg feTile>
|       <svg feTurbulence>
|       <svg foreignObject>
|       <svg glyphRef>
|       <svg linearGradient>
|       <svg radialGradient>
|       <svg textPath>

Test 7 of 9 in resources/tests11.dat failed. Input:
<!DOCTYPE html><BODY><SVG><ALTGLYPH /><ALTGLYPHDEF /><ALTGLYPHITEM /><ANIMATECOLOR /><ANIMATEMOTION /><ANIMATETRANSFORM /><CLIPPATH /><FEBLEND /><FECOLORMATRIX /><FECOMPONENTTRANSFER /><FECOMPOSITE /><FECONVOLVEMATRIX /><FEDIFFUSELIGHTING /><FEDISPLACEMENTMAP /><FEDISTANTLIGHT /><FEFLOOD /><FEFUNCA /><FEFUNCB /><FEFUNCG /><FEFUNCR /><FEGAUSSIANBLUR /><FEIMAGE /><FEMERGE /><FEMERGENODE /><FEMORPHOLOGY /><FEOFFSET /><FEPOINTLIGHT /><FESPECULARLIGHTING /><FESPOTLIGHT /><FETILE /><FETURBULENCE /><FOREIGNOBJECT /><GLYPHREF /><LINEARGRADIENT /><RADIALGRADIENT /><TEXTPATH /></SVG>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|       <altglyph>
|         <altglyphdef>
|           <altglyphitem>
|             <animatecolor>
|               <animatemotion>
|                 <animatetransform>
|                   <clippath>
|                     <feblend>
|                       <fecolormatrix>
|                         <fecomponenttransfer>
|                           <fecomposite>
|                             <feconvolvematrix>
|                               <fediffuselighting>
|                                 <fedisplacementmap>
|                                   <fedistantlight>
|                                     <feflood>
|                                       <fefunca>
|                                         <fefuncb>
|                                           <fefuncg>
|                                             <fefuncr>
|                                               <fegaussianblur>
|                                                 <feimage>
|                                                   <femerge>
|                                                     <femergenode>
|                                                       <femorphology>
|                                                         <feoffset>
|                                                           <fepointlight>
|                                                             <fespecularlighting>
|                                                               <fespotlight>
|                                                                 <fetile>
|                                                                   <feturbulence>
|                                                                     <foreignobject>
|                                                                       <glyphref>
|                                                                         <lineargradient>
|                                                                           <radialgradient>
|                                                                             <textpath>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg altGlyph>
|       <svg altGlyphDef>
|       <svg altGlyphItem>
|       <svg animateColor>
|       <svg animateMotion>
|       <svg animateTransform>
|       <svg clipPath>
|       <svg feBlend>
|       <svg feColorMatrix>
|       <svg feComponentTransfer>
|       <svg feComposite>
|       <svg feConvolveMatrix>
|       <svg feDiffuseLighting>
|       <svg feDisplacementMap>
|       <svg feDistantLight>
|       <svg feFlood>
|       <svg feFuncA>
|       <svg feFuncB>
|       <svg feFuncG>
|       <svg feFuncR>
|       <svg feGaussianBlur>
|       <svg feImage>
|       <svg feMerge>
|       <svg feMergeNode>
|       <svg feMorphology>
|       <svg feOffset>
|       <svg fePointLight>
|       <svg feSpecularLighting>
|       <svg feSpotLight>
|       <svg feTile>
|       <svg feTurbulence>
|       <svg foreignObject>
|       <svg glyphRef>
|       <svg linearGradient>
|       <svg radialGradient>
|       <svg textPath>

Test 8 of 9 in resources/tests11.dat failed. Input:
<!DOCTYPE html><body><math><altGlyph /><altGlyphDef /><altGlyphItem /><animateColor /><animateMotion /><animateTransform /><clipPath /><feBlend /><feColorMatrix /><feComponentTransfer /><feComposite /><feConvolveMatrix /><feDiffuseLighting /><feDisplacementMap /><feDistantLight /><feFlood /><feFuncA /><feFuncB /><feFuncG /><feFuncR /><feGaussianBlur /><feImage /><feMerge /><feMergeNode /><feMorphology /><feOffset /><fePointLight /><feSpecularLighting /><feSpotLight /><feTile /><feTurbulence /><foreignObject /><glyphRef /><linearGradient /><radialGradient /><textPath /></math>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math>
|       <altglyph>
|         <altglyphdef>
|           <altglyphitem>
|             <animatecolor>
|               <animatemotion>
|                 <animatetransform>
|                   <clippath>
|                     <feblend>
|                       <fecolormatrix>
|                         <fecomponenttransfer>
|                           <fecomposite>
|                             <feconvolvematrix>
|                               <fediffuselighting>
|                                 <fedisplacementmap>
|                                   <fedistantlight>
|                                     <feflood>
|                                       <fefunca>
|                                         <fefuncb>
|                                           <fefuncg>
|                                             <fefuncr>
|                                               <fegaussianblur>
|                                                 <feimage>
|                                                   <femerge>
|                                                     <femergenode>
|                                                       <femorphology>
|                                                         <feoffset>
|                                                           <fepointlight>
|                                                             <fespecularlighting>
|                                                               <fespotlight>
|                                                                 <fetile>
|                                                                   <feturbulence>
|                                                                     <foreignobject>
|                                                                       <glyphref>
|                                                                         <lineargradient>
|                                                                           <radialgradient>
|                                                                             <textpath>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math altglyph>
|       <math altglyphdef>
|       <math altglyphitem>
|       <math animatecolor>
|       <math animatemotion>
|       <math animatetransform>
|       <math clippath>
|       <math feblend>
|       <math fecolormatrix>
|       <math fecomponenttransfer>
|       <math fecomposite>
|       <math feconvolvematrix>
|       <math fediffuselighting>
|       <math fedisplacementmap>
|       <math fedistantlight>
|       <math feflood>
|       <math fefunca>
|       <math fefuncb>
|       <math fefuncg>
|       <math fefuncr>
|       <math fegaussianblur>
|       <math feimage>
|       <math femerge>
|       <math femergenode>
|       <math femorphology>
|       <math feoffset>
|       <math fepointlight>
|       <math fespecularlighting>
|       <math fespotlight>
|       <math fetile>
|       <math feturbulence>
|       <math foreignobject>
|       <math glyphref>
|       <math lineargradient>
|       <math radialgradient>
|       <math textpath>

Test 9 of 9 in resources/tests11.dat failed. Input:
<!DOCTYPE html><body><svg><solidColor /></svg>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg>
|       <solidcolor>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg solidcolor>
resources/tests12.dat:
1
2

Test 1 of 2 in resources/tests12.dat failed. Input:
<!DOCTYPE html><body><p>foo<math><mtext><i>baz</i></mtext><annotation-xml><svg><desc><b>eggs</b></desc><g><foreignObject><P>spam<TABLE><tr><td><img></td></table></foreignObject></g><g>quux</g></svg></annotation-xml></math>bar
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       "foo"
|       <math>
|         <mtext>
|           <i>
|             "baz"
|         <annotation-xml>
|           <svg>
|             <desc>
|               <b>
|                 "eggs"
|             <g>
|               <foreignobject>
|     <p>
|       "spam"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <img>
|     <g>
|       "quux"
|     "bar"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       "foo"
|       <math math>
|         <math mtext>
|           <i>
|             "baz"
|         <math annotation-xml>
|           <svg svg>
|             <svg desc>
|               <b>
|                 "eggs"
|             <svg g>
|               <svg foreignObject>
|                 <p>
|                   "spam"
|                 <table>
|                   <tbody>
|                     <tr>
|                       <td>
|                         <img>
|             <svg g>
|               "quux"
|       "bar"

Test 2 of 2 in resources/tests12.dat failed. Input:
<!DOCTYPE html><body>foo<math><mtext><i>baz</i></mtext><annotation-xml><svg><desc><b>eggs</b></desc><g><foreignObject><P>spam<TABLE><tr><td><img></td></table></foreignObject></g><g>quux</g></svg></annotation-xml></math>bar
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "foo"
|     <math>
|       <mtext>
|         <i>
|           "baz"
|       <annotation-xml>
|         <svg>
|           <desc>
|             <b>
|               "eggs"
|           <g>
|             <foreignobject>
|               <p>
|                 "spam"
|               <table>
|                 <tbody>
|                   <tr>
|                     <td>
|                       <img>
|           <g>
|             "quux"
|     "bar"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "foo"
|     <math math>
|       <math mtext>
|         <i>
|           "baz"
|       <math annotation-xml>
|         <svg svg>
|           <svg desc>
|             <b>
|               "eggs"
|           <svg g>
|             <svg foreignObject>
|               <p>
|                 "spam"
|               <table>
|                 <tbody>
|                   <tr>
|                     <td>
|                       <img>
|           <svg g>
|             "quux"
|     "bar"
resources/tests14.dat: PASS

resources/tests15.dat:
7
8
9
10
11

Test 7 of 14 in resources/tests15.dat failed. Input:
<!doctype html><table> X<meta></table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <meta>
|     <table>
|       " X"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     " X"
|     <meta>
|     <table>

Test 8 of 14 in resources/tests15.dat failed. Input:
<!doctype html><table> x</table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       " x"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     " x"
|     <table>

Test 9 of 14 in resources/tests15.dat failed. Input:
<!doctype html><table> x </table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       " x "
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     " x "
|     <table>

Test 10 of 14 in resources/tests15.dat failed. Input:
<!doctype html><table><tr> x</table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           " x"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     " x"
|     <table>
|       <tbody>
|         <tr>

Test 11 of 14 in resources/tests15.dat failed. Input:
<!doctype html><table>X<style> <tr>x </style> </table>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       "X"
|       <style>
|         " <tr>x "
|       " "
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "X"
|     <table>
|       <style>
|         " <tr>x "
|       " "
resources/tests16.dat: PASS

resources/webkit01.dat: PASS

resources/doctype01.dat: PASS

resources/scriptdata01.dat: PASS

resources/entities01.dat:
2
5

Test 2 of 68 in resources/entities01.dat failed. Input:
FOO&gtBAR
Got:
| <html>
|   <head>
|   <body>
|     "FOO&gtBAR"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO>BAR"

Test 5 of 68 in resources/entities01.dat failed. Input:
I'm &notit; I tell you
Got:
| <html>
|   <head>
|   <body>
|     "I'm &notit; I tell you"
Expected:
| <html>
|   <head>
|   <body>
|     "I'm it; I tell you"
resources/entities02.dat: PASS

resources/comments01.dat: PASS

resources/adoption01.dat:
3

Test 3 of 9 in resources/adoption01.dat failed. Input:
<a>1<button>2</a>3</button>
Got:
| <html>
|   <head>
|   <body>
|     <a>
|       "1"
|       <button>
|         "23"
Expected:
| <html>
|   <head>
|   <body>
|     <a>
|       "1"
|       <button>
|         "2"
|     "3"
resources/inbody01.dat: PASS

resources/isindex.dat: PASS

resources/tables01.dat:
4

Test 4 of 15 in resources/tables01.dat failed. Input:
<table><colgroup></html>foo
Got:
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>
|       "foo"
Expected:
| <html>
|   <head>
|   <body>
|     "foo"
|     <table>
|       <colgroup>
#EOF
